+++ activerecord/lib/active_record/attribute_set.rb
@@ -3,7 +3,7 @@ require 'active_record/attribute_set/yaml_encoder'
 
 module ActiveRecord
   class AttributeSet # :nodoc:
-    delegate :each_value, to: :attributes
+    delegate :each_value, :fetch, to: :attributes
 
     def initialize(attributes)
       @attributes = attributes